Feat: Heatmap generation
Currently have code setup to generate a folder of heatmaps
based on a text file I use to capture stdout. This is a temporary
setup until the benchmark is complete, then I will retrieve the data
using json to be consistent with the other benchmarks.

I also modularized the code a bit more.